--- bouncer/src/repo/manipulator/modelconvertor/import/ifcHelper/repo_ifc_helper_common.h
@@ -1,34 +1,0 @@
-/**
-*  Copyright (C) 2021 3D Repo Ltd
-*
-*  This program is free software: you can redistribute it and/or modify
-*  it under the terms of the GNU Affero General Public License as
-*  published by the Free Software Foundation, either version 3 of the
-*  License, or (at your option) any later version.
-*
-*  This program is distributed in the hope that it will be useful,
-*  but WITHOUT ANY WARRANTY; without even the implied warranty of
-*  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-*  GNU Affero General Public License for more details.
-*
-*  You should have received a copy of the GNU Affero General Public License
-*  along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#pragma once
-#include <string>
-#include <ifcparse/IfcFile.h>
-
-enum IfcSchemaVersion { IFC2x3, IFC4, UNKNOWN };
-
-static IfcSchemaVersion getIFCSchema(const std::string &file) {
-	IfcParse::IfcFile ifcFile = file;
-	if (ifcFile.schema()->name() == "IFC2X3") {
-		return IfcSchemaVersion::IFC2x3;
-	}
-	else if (ifcFile.schema()->name() == "IFC4") {
-		return IfcSchemaVersion::IFC4;
-	}
-
-	return IfcSchemaVersion::UNKNOWN;
-}
